+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Develop
 
 ### Added
+- Add conversion from bigwig to bedgraph and wig formats
+- Ability to calculate queries at lower resolutions with bwtools
 - Ability to downsample aligned data based on group minimum
 - Gini coefficient calculations for NETseq data
 - Pseudoalignment with Kallisto for stranded data